pyproject.toml
[tool.poetry]
name = "indoor-highlights"
version = "0.1.0"
description = ""
authors = ["Shane Whelan <shanewhelan89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
moviepy = "==2.0.0.dev2"
python-dotenv = "^1.0.1"
ipdb = "^0.13.13"
google-api-python-client = "^2.150.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
